cambios antes de pasar a main

+++ src/main/java/com/asier/SistemaReservas/seats/domain/DTO/SeatDTO.java
@@ -22,5 +22,5 @@ public class SeatDTO {
     private SeatType seatType;
     private BigDecimal costPerSeat;
     private Boolean available = true;
-    private Integer flightId;
+    private Long flightId;
 }
